+++ api_rest_taller/src/main/java/com/example/demo/Interface.java
@@ -2,5 +2,5 @@ package com.example.demo;
 
 public interface Interface {
 	public void altamatricula(String entrada, String matricula);
-	public int consultamatricula(String matricula);
+	public double consultamatricula(String matricula);
 }
